Clear global temp tables before populating them with package procedure parameters
The global temp tables that hold collection parameters for packages need
to be cleared before they are populated rather than only after the
procedure or function completes. If something goes wrong in the execution
of the Oracle procedure, the global temp table for the parameter will not get
cleared after the call. This leaves behind garbage in the temp table that
will make the subsequent call to the same procedure wrong.

By putting tmp_table.delete right before populating tmp_table there is a
guarantee the no parameters from a previous function call will get passed
on to the current function call.